refactor(deprecate): rename to routine + messaging
- rename `callables.py` -> `routine.py` (function/method decorators; singular like `proxy`/`module`, no `callable` builtin clash)
- rename `messages.py` -> `messaging.py`
- rename the mirror unit tests `test_callables.py` -> `test_routine.py` and `test_messages.py` -> `test_messaging.py`
- repoint all imports, the `deprecation.py` shim, the api-reference autodoc paths, and the CONTRIBUTING/AGENTS structure trees; public API names unchanged

---
